refactor: create common folder and move common files on it
